+++ kolobok-annotations/src/main/java/org/kolobok/annotation/LogContext.java
@@ -1,0 +1,11 @@
+package org.kolobok.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.CLASS)
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface LogContext {
+}