+++ samples/spring-data-jpa-gradle/src/main/java/com/example/kolobok/PersonRepository.java
@@ -1,11 +1,11 @@
 package com.example.kolobok;
 
 import org.kolobok.annotation.FindWithOptionalParams;
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PersonRepository extends CrudRepository<Person, Long> {
+public interface PersonRepository extends JpaRepository<Person, Long> {
     @FindWithOptionalParams
     List<Person> findByFirstNameAndLastName(String firstName, String lastName);